pyproject.toml
[project]
name = "growth-analytics-pipeline"
version = "0.1.0"
description = "SaaS PLG Growth Analytics Pipeline with Delta Lake and Airflow"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]

dependencies = [
    # Core Data Engineering
    "pyspark>=3.5.0,<4.0.0",  # Pin to 3.5.x for compatibility with delta-spark
    "delta-spark>=3.2.0,<4.0.0",  # Pin to 3.2.x for compatibility with PySpark 3.5.x

    # Airflow
    "apache-airflow>=2.8.0",

    # Data Generation & Processing
    "faker>=22.0.0",
    "pandas>=2.1.4",
    "numpy>=1.26.3",

    # Jupyter & Analysis
    "jupyter>=1.0.0",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.11",
    "black>=23.12.1",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spark", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "Q"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["spark/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=spark --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = false
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
strict_equality = false
# Start with lenient settings, gradually increase strictness
ignore_missing_imports = true
exclude = [
    "^.venv/",
    "^build/",
    "^dist/",
]
